score is now a 16-bit integer

+++ ByggmesterMartha/ByggmesterMartha/inc/game.h
@@ -12,7 +12,7 @@
 #include "menu.h"
 #include "statemachine.h"
 
-volatile uint8_t score_count;
+volatile uint16_t score_count;
 uint8_t game_over;
 
 void game_init();
